+++ com/back/Main.java
@@ -1,0 +1,27 @@
+package com.back;
+
+import java.util.Scanner;
+
+public class Main {
+    public static void main(String[] args) {
+        Scanner sc = new Scanner(System.in);
+
+        System.out.println("== 명언 앱 ==");
+
+        while (true) {
+            System.out.print("명령) ");
+            String command = sc.nextLine().trim();
+
+            if (command.equals("등록")) {
+                System.out.print("명언 : ");
+                sc.nextLine();
+                System.out.print("작가 : ");
+                sc.nextLine();
+            }
+            else if (command.equals("종료")) {
+                System.out.println("앱을 종료합니다.");
+                break;
+            }
+        }
+    }
+}
