--- src/main/java/com/back/Main.java
@@ -1,13 +1,0 @@
-package src.main.java.com.back;
-
-import java.util.Scanner;
-
-public class Main {
-    public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-
-        System.out.println("== 명언 앱 ==");
-        System.out.print("명령) ");
-        sc.nextLine();
-    }
-}
